requirements.txt
# Core FastAPI and web framework
fastapi==0.115.12
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Pydantic for data validation
pydantic==2.11.3
pydantic_core==2.33.1
pydantic-settings==2.9.1

# LangChain and LangGraph for AI workflow
langchain==0.3.24
langchain-core==0.3.59
langchain-openai==0.3.14
langgraph==0.3.31
langsmith==0.3.33
langchain-unstructured==0.1.5

# HTML parsing and processing
beautifulsoup4==4.12.2
lxml==4.9.3

# HTTP client for external requests
httpx==0.28.1
httpx-sse==0.4.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Optional: For enhanced AI capabilities
openai==1.75.0
tiktoken==0.9.0
anyio==4.9.0
langchain-community==0.3.22
unstructured-client==0.25.9
